+++ src/pyi_updater/package_handler/__init__.py
@@ -12,11 +12,10 @@ except ImportError:
 from jms_utils.paths import ChDir
 
 from pyi_updater.package_handler.package import Package, Patch
+from pyi_updater.package_handler.utils import migrate
 from pyi_updater.utils import (get_package_hashes as gph,
                                remove_dot_files,
                                EasyAccessDict,
-                               version_string_to_tuple,
-                               version_tuple_to_string,
                                )
 
 log = logging.getLogger(__name__)
@@ -57,12 +56,21 @@ class PackageHandler(object):
             self.files_dir = os.path.join(self.data_dir, u'files')
             self.deploy_dir = os.path.join(self.data_dir, u'deploy')
             self.new_dir = os.path.join(self.data_dir, u'new')
+            self.config_dir = os.path.join(os.path.dirname(self.data_dir),
+                                           u'.pyiupdater')
+            self.config_file = os.path.join(self.config_dir, 'data.json')
+            self.config = None
         else:
             log.debug('DEV_DATA_DIR is None. Setup Failed')
 
         self.update_url = obj.config.get(u'UPDATE_URL')
 
         self.json_data = None
+        if self.data_dir is not None and \
+                os.path.exists(self.files_dir) is True:
+            migrate(self.data_dir)
+        if self.data_dir is not None:
+            self.setup()
 
     def setup(self):
         """Creates all needed working directories & loads version file.
@@ -71,6 +79,7 @@ class PackageHandler(object):
         """
         self._setup_work_dirs()
         self.json_data = self._load_version_file()
+        self.config = self._load_config()
 
     def process_packages(self):
         """Gets a list of updates to process.  Adds the name of an
@@ -80,17 +89,18 @@ class PackageHandler(object):
 
         Proxy method for :meth:`_get_package_list`,
         :meth:`_make_patches`, :meth:`_add_patches_to_packages`,
-        :meth:`_setup_file_dirs`, :meth:`_update_version_file`,
+        :meth:`_update_version_file`,
         :meth:`_write_json_to_file` & :meth:`_move_packages`.
         """
         package_manifest, patch_manifest = self._get_package_list()
         patches = self._make_patches(patch_manifest)
+        self._cleanup(patch_manifest)
         package_manifest = self._add_patches_to_packages(package_manifest,
                                                          patches)
-        self._setup_file_dirs(package_manifest)
         self.json_data = self._update_version_file(self.json_data,
                                                    package_manifest)
         self._write_json_to_file(self.json_data)
+        self._write_config_to_file(self.config)
         self._move_packages(package_manifest)
 
     def _setup_work_dirs(self):
@@ -103,7 +113,9 @@ class PackageHandler(object):
         #
         # This is non destructive
         log.debug(u'Setting up work dirs...')
-        dirs = [self.data_dir, self.new_dir, self.deploy_dir, self.files_dir]
+        dirs = [self.data_dir, self.new_dir,
+                self.deploy_dir, self.files_dir,
+                self.config_dir]
         for d in dirs:
             if not os.path.exists(d):
                 os.mkdir(d)
@@ -138,6 +150,22 @@ class PackageHandler(object):
         log.debug(u'Loaded version file')
         return json_data
 
+    def _load_config(self):
+        config = None
+        log.debug(u'Looking for config file...')
+        if os.path.exists(self.config_file):
+            try:
+                with open(self.config_file, u'r') as f:
+                    config = json.loads(f.read())
+            except Exception as err:
+                log.debug(str(err))
+
+        if config is None:
+            config = {
+                u'patches': {}
+                }
+        return config
+
     def _get_package_list(self, ignore_errors=True):
         # Adds compatible packages to internal package manifest
         # for futher processing
@@ -146,9 +174,9 @@ class PackageHandler(object):
         log.debug(u'Getting package list')
         # Clears manifest if sign updates runs more the once without
         # app being restarted
-        package_manifest = []
-        patch_manifest = []
-        bad_packages = []
+        package_manifest = list()
+        patch_manifest = list()
+        bad_packages = list()
         with ChDir(self.new_dir):
             # Getting a list of all files in the new dir
             packages = os.listdir(os.getcwd())
@@ -169,14 +197,17 @@ class PackageHandler(object):
                                                         package)
 
                 package_manifest.append(package)
+                self.config = self._add_package_to_config(package,
+                                                          self.config)
 
                 if self.patch_support:
                     # Will check if source file for patch exists
                     # if so will return the path and number of patch
                     # to create. If missing source file None returned
-                    path = self._check_make_patch(package.name,
-                                                  package.version,
-                                                  package.platform)
+                    path = self._check_make_patch(self.json_data,
+                                                  package.name,
+                                                  package.platform,
+                                                  )
                     if path is not None:
                         log.debug(u'Found source file to create patch')
                         patch_name = package.name + u'-' + package.platform
@@ -201,6 +232,33 @@ class PackageHandler(object):
 
         return package_manifest, patch_manifest
 
+    def _add_package_to_config(self, p, data):
+        if 'package' not in data.keys():
+            data['package'] = {}
+            log.debug('Initilizing config for packages')
+        # First package with current name so add platform and version
+        if p.name not in data['package'].keys():
+            data['package'][p.name] = {p.platform: p.version}
+            log.debug('Adding new package to config')
+        else:
+            # Adding platform and version
+            if p.platform not in data['package'][p.name].keys():
+                data['package'][p.name][p.platform] = p.version
+                log.debug('Adding new arch to package-config')
+            else:
+                # Getting current version for platform
+                value = data['package'][p.name][p.platform]
+                # Updating version if applicable
+                if p.version > value:
+                    log.debug('Adding new version to package-config')
+                    data['package'][p.name][p.platform] = p.version
+        return data
+
+    def _cleanup(self, patch_manifest):
+        for p in patch_manifest:
+            if os.path.exists(p[u'src']):
+                os.remove(p[u'src'])
+
     def _make_patches(self, patch_manifest):
         # ToDo: Since not packing as an executable test
         #       to see if it multiprocessing works on windows
@@ -216,7 +274,7 @@ class PackageHandler(object):
         # else:
         cpu_count = multiprocessing.cpu_count() * 2
         pool = multiprocessing.Pool(processes=cpu_count)
-        pool_output = pool.map(_make_patch, patch_manifest)
+        pool_output = pool.map(self._make_patch, patch_manifest)
         return pool_output
 
     def _add_patches_to_packages(self, package_manifest, patches):
@@ -237,19 +295,6 @@ class PackageHandler(object):
                     else:
                         log.debug('No patch match found')
         return package_manifest
-
-    def _setup_file_dirs(self, package_manifest):
-        log.debug(u'Setting up directories for file updates')
-        for p in package_manifest:
-            package_dir = os.path.join(self.files_dir, p.name)
-            package_version_path = os.path.join(package_dir, p.version)
-            p.version_path = package_version_path
-            # Creating a dir for specific package
-            if not os.path.exists(package_dir):
-                os.mkdir(package_dir)
-            # Creating a dir for specific version of above package
-            if not os.path.exists(package_version_path):
-                os.mkdir(package_version_path)
 
     def _update_version_file(self, json_data, package_manifest):
         # Updates version file with package meta-data
@@ -299,6 +344,11 @@ class PackageHandler(object):
         with open(os.path.join(self.data_dir, u'version.json'), u'w') as f:
             f.write(json.dumps(json_data, sort_keys=True, indent=4))
 
+    def _write_config_to_file(self, json_data):
+        log.debug(u'Writing config data to file')
+        with open(self.config_file, u'w') as f:
+            f.write(json.dumps(json_data, sort_keys=True, indent=2))
+
     # ToDo: Explain whats going on below &/or clean up
     def _move_packages(self, package_manifest):
         # Moves all packages to their destination folder.
@@ -309,24 +359,21 @@ class PackageHandler(object):
         log.debug(u'Moving packages to deploy folder')
         for p in package_manifest:
             patch = p.patch_info.get(u'patch_name', None)
-            version_path = p.version_path
             with ChDir(self.new_dir):
                 if patch:
-                    shutil.copy(patch, self.deploy_dir)
+                    shutil.move(patch, self.deploy_dir)
                     log.debug(u'Copying {} to {}'.format(patch,
                               self.deploy_dir))
-                    shutil.move(patch, version_path)
-                    log.debug(u'Moving {} to {}'.format(patch, version_path))
 
                 shutil.copy(p.filename, self.deploy_dir)
                 log.debug(u'Copying {} to {}'.format(p.filename,
                           self.deploy_dir))
 
-                if os.path.exists(os.path.join(version_path, p.filename)):
-                    os.remove(os.path.join(version_path, p.filename))
-                shutil.move(p.filename, version_path)
+                if os.path.exists(os.path.join(self.files_dir, p.filename)):
+                    os.remove(os.path.join(self.files_dir, p.filename))
+                shutil.move(p.filename, self.files_dir)
                 log.debug(u'Moving {} to {}'.format(p.filename,
-                          version_path))
+                          self.files_dir))
 
     def _update_file_list(self, json_data, package_info):
         files = json_data[u'updates']
@@ -343,102 +390,69 @@ class PackageHandler(object):
             json_data[u'latest'][package_info.name] = {}
         return json_data
 
-    def _check_make_patch(self, name, version_str, platform):
+    def _check_make_patch(self, json_data, name, platform):
         # Check to see if previous version is available to
         # make patch updates
         # Also calculates patch number
         log.debug('Checking if patch creation is possible')
-        version = version_string_to_tuple(version_str)
         if bsdiff4 is None:
             return None
         src_file_path = None
-        highest_version_str = None
-        data_dir = os.path.join(self.files_dir, name)
-        if os.path.exists(data_dir):
-            with ChDir(data_dir):
-                # getting a list of all version folders
-                # for current app
-                version_dirs = os.listdir(os.getcwd())
-                version_dirs = remove_dot_files(version_dirs)
+        if os.path.exists(self.files_dir):
+            with ChDir(self.files_dir):
+                files = os.listdir(os.getcwd())
 
-                # Changing version strings into tuples
-                fixed_version_dirs = []
-                for v in version_dirs:
-                    fixed_version_dirs.append(version_string_to_tuple(v))
-
-                # Can't make a patch if no source file to work with
-                if len(fixed_version_dirs) < 1:
-                    return None
-
-                # highest_version = sorted(fixed_version_dirs)[-1]
-                # highest_version_str = version_tuple_to_string(highest_version)
-
-                # # ToDo: Not sure if this is need.
-                # #       Will research later.
-                # if highest_version > version:
-                #     return None
-
-            found = False
-            versions_hi_to_low = sorted(fixed_version_dirs, reverse=True)
-            for v in versions_hi_to_low:
-                if found is True:
-                    break
-                files = os.listdir(os.path.join(data_dir,
-                                   version_tuple_to_string(v)))
-                for f in files:
-                    if platform in f:
-                        found = True
-                        highest_version_str = version_tuple_to_string(v)
-                        break
-
-            if highest_version_str is None:
+            files = remove_dot_files(files)
+            if len(files) == 0:
                 return None
-            # Chainging into directory to get absolute path of
-            # source file for patch creation later
-            target_dir = os.path.join(data_dir, highest_version_str)
-            with ChDir(target_dir):
-                files = remove_dot_files(os.listdir(os.getcwd()))
-                log.debug('Files in {}:\n{}'.format(target_dir, files))
-                # If somehow the source file got deleted
-                # just return
-                if len(files) == 0:
-                    return None
-                # Getting the correct source matches
-                # destination platform
-                for f in files:
-                    if highest_version_str in f and platform in f:
-                        log.debug('Found src file to create patch')
-                        log.debug('Src file: {}'.format(f))
-                        src_file_path = os.path.abspath(f)
-                        break
-            # if our list gets exhausted before finding
-            # source file then just return None
-            if src_file_path is None:
+            try:
+                latest = json_data[u'latest'][name][platform]
+            except KeyError:
                 return None
-            return src_file_path, len(fixed_version_dirs)
+            try:
+                l_plat = json_data[u'updates'][name][latest][platform]
+                filename = l_plat[u'filename']
+            except:
+                return None
+            src_file_path = os.path.join(self.files_dir, filename)
 
+            try:
+                patch_num = self.config[u'patches'][name]
+                self.config[u'patches'][name] += 1
+            except KeyError:
+                try:
+                    patch_num = self.config[u'boot_strap']
+                except KeyError:
+                    patch_num = 100
+                if u'patches' not in self.config.keys():
+                    self.config[u'patches'] = {}
+                if name not in self.config[u'patches'].keys():
+                    self.config[u'patches'][name] = patch_num
+            num = patch_num + 1
+            return src_file_path, num
+        return None
 
-def _make_patch(patch_info):
-    # Does with the name implies. Used with multiprocessing
-    patch = Patch(patch_info)
-    patch_name = patch_info[u'patch_name']
-    dst_path = patch_info[u'dst']
-    patch_number = patch_info[u'patch_num']
-    src_path = patch_info[u'src']
-    patch_name += u'-' + str(patch_number)
-    # Updating with full name - number included
-    patch.patch_name = patch_name
-    if not os.path.exists(src_path):
-        log.debug('Src file does not exist to create patch')
+    def _make_patch(self, patch_info):
+        # Does with the name implies. Used with multiprocessing
+        patch = Patch(patch_info)
+        patch_name = patch_info[u'patch_name']
+        dst_path = patch_info[u'dst']
+        patch_number = patch_info[u'patch_num']
+        src_path = patch_info[u'src']
+        patch_name += u'-' + str(patch_number)
+        # Updating with full name - number included
+        patch.patch_name = patch_name
+        if not os.path.exists(src_path):
+            log.debug('Src file does not exist to create patch')
 
-    else:
-        print(u"Making patch... {}".format(os.path.basename(patch_name)))
-        log.debug(u'Patch source path:{}'.format(src_path))
-        log.debug(u'Patch destination path: {}'.format(dst_path))
-        if patch.ready is True:
-            log.debug(u'Creating patch')
-            bsdiff4.file_diff(src_path, patch.dst_path, patch.patch_name)
-            log.debug(u'Done creating patch')
         else:
-            log.debug('Missing patch attr')
-    return patch
+            print(u"Making patch... {}".format(os.path.basename(patch_name)))
+            log.debug(u'Patch source path:{}'.format(src_path))
+            log.debug(u'Patch destination path: {}'.format(dst_path))
+            if patch.ready is True:
+                log.debug(u'Creating patch')
+                bsdiff4.file_diff(src_path, patch.dst_path, patch.patch_name)
+                log.debug(u'Done creating patch')
+            else:
+                log.debug('Missing patch attr')
+        return patch